+++ source/plugins/plgcommon/plghost_interface.h
@@ -101,13 +101,13 @@ struct ipcr // reader
     ipcr(const void *dd, int sz) :d((const char *)dd), sz(sz) {}
 
     const data_header_s&header() const { return *(data_header_s *)d; }
-    template<typename T> const T& get() { int rptr = ptr; ptr += sizeof(T); ASSERT(ptr <= sz); return *(T *)(d+rptr); }
+    template<typename T> const T& get() { int rptr = ptr; ptr += sizeof(T); ASSERT(ptr <= sz); return *(T *)(d + ptrdiff_t(rptr)); }
     const void *read_data(int rsz)
     {
         int rptr = ptr;
         ptr += rsz;
         if (ASSERT(ptr <= sz))
-            return d + rptr;
+            return d + ptrdiff_t(rptr);
         return nullptr;
     }
     const void *get_data(int &rsz)
@@ -119,15 +119,15 @@ struct ipcr // reader
     const char *readastr(int &len)
     {
         len = get<unsigned short>();
-        const char *s = d + ptr;
+        const char *s = d + ptrdiff_t(ptr);
         ptr += len;
         ASSERT(ptr <= sz);
         return s;
     }
 
 #ifdef STRTYPE
-    STRTYPE(char) getastr() { int l = get<unsigned short>(); const char *s = d + ptr; ptr += l; ASSERT(ptr <= sz); return MAKESTRTYPE( char, s, l ); }
-    STRTYPE(wchar_t) getwstr() { int l = get<unsigned short>(); const wchar_t *s = (const wchar_t *)(d + ptr); ptr += l * sizeof(wchar_t); ASSERT(ptr <= sz); return MAKESTRTYPE( wchar_t, s, l ); }
+    STRTYPE(char) getastr() { int l = get<unsigned short>(); const char *s = d + ptrdiff_t(ptr); ptr += l; ASSERT(ptr <= sz); return MAKESTRTYPE( char, s, l ); }
+    STRTYPE(wchar_t) getwstr() { int l = get<unsigned short>(); const wchar_t *s = (const wchar_t *)(d + ptrdiff_t(ptr)); ptr += l * sizeof(wchar_t); ASSERT(ptr <= sz); return MAKESTRTYPE( wchar_t, s, l ); }
 #endif
 };
